==> notebooks/12_frontend.ipynb <==
import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))
import src.config as config

from src.inference import fetch_hourly_rides, fetch_predictions

df1 = fetch_hourly_rides(12)
df2 = fetch_predictions(12)
print("mnbdsjhv")
print(df1)
print("sdwdsd")
print(df2)

import pandas as pd

merged_df = pd.merge(df1, df2, on=["pickup_location_id", "pickup_hour"])

print(merged_df)

import geopandas as gpd

# Replace with the actual path to your downloaded .shp file
shape_url = "https://data.jerseycitynj.gov/api/explore/v2.1/catalog/datasets/citi-bike-locations-phase-1-system-map-3/exports/shp?lang=en&timezone=America%2FNew_York"
gdf = gpd.read_file(shape_url)

print(gdf.columns.tolist())
